refactor: Expressions 순수 함수를 통해 evaluate 하도록 수정
