+++ src/main/kotlin/next/step/calculator/service/Calculator.kt
@@ -1,7 +1,5 @@
 package next.step.calculator.service
 
-import next.step.calculator.domain.Expressions
-
 object Calculator {
-    fun calculate(command: String): Int = Expressions.of(Parser.parse(command)).evaluate()
+    fun calculate(command: String): Int = Expressions.evaluate(Parser.parse(command))
 }
