+++ src/main/kotlin/next/step/calculator/service/Expressions.kt
@@ -1,16 +1,12 @@
-package next.step.calculator.domain
+package next.step.calculator.service
 
-class Expressions(private val expressions: List<String>) {
-    companion object {
-        fun of(expressions: List<String>): Expressions {
-            require(isEnoughCount(expressions)) { "expression 들은 홀수 개여야 합니다." }
-            return Expressions(expressions)
-        }
+import next.step.calculator.domain.InputNumber
+import next.step.calculator.domain.Operator
 
-        private fun isEnoughCount(expressions: List<String>) = expressions.size % 2 == 1
-    }
+object Expressions {
+    fun evaluate(expressions: List<String>): Int {
+        require(isEnoughCount(expressions)) { "expression 들은 홀수 개여야 합니다." }
 
-    fun evaluate(): Int {
         var result = InputNumber(0)
         var operator = Operator.PLUS
         for ((index, expression) in expressions.withIndex()) {
@@ -23,5 +19,6 @@ class Expressions(private val expressions: List<String>) {
         return result.toInt()
     }
 
+    private fun isEnoughCount(expressions: List<String>) = expressions.size % 2 == 1
     private fun isNumberIndex(index: Int) = index % 2 == 0
 }
